4주차 과제

+++ src/main/java/com/likelion/lionshop/controller/OrderController.java
@@ -8,6 +8,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -22,7 +24,12 @@ public class OrderController {
 
     // 1. 주문을 생성하는 컨트롤러를 만듭니다. 이때 return 값은 "주문 생성하기"입니다. -> 주문은 리스트 형태로 요청을 보내주세요!
     @PostMapping(value = "")
-    public ResponseEntity<List<OrderResponseDto>> createOrders(@RequestBody List<CreateOrderRequestDto> createOrderRequestDtos) { //List형태로 주문을 받을 수 있게 수정
+    public ResponseEntity<List<OrderResponseDto>> createOrders(@RequestBody List<CreateOrderRequestDto> createOrderRequestDtos,
+                                                               @AuthenticationPrincipal UserDetails userDetails) { //List형태로 주문을 받을 수 있게 수정
+        if(userDetails == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
         List<OrderResponseDto> orderResponseDtos = orderService.createOrders(createOrderRequestDtos);
         return new ResponseEntity<>(orderResponseDtos, HttpStatus.CREATED);
     }
@@ -30,7 +37,12 @@ public class OrderController {
 
     // 2. 주문을 가져오는 컨트롤러를 만듭니다. 이때 return 값은 "주문 가져오기"입니다.
     @GetMapping(value = "/{orderId}")
-    public ResponseEntity<OrderResponseDto> getOrder(@PathVariable("orderId") Long id) {
+    public ResponseEntity<OrderResponseDto> getOrder(@PathVariable("orderId") Long id,
+                                                     @AuthenticationPrincipal UserDetails userDetails) {
+        if(userDetails == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
         OrderResponseDto orderResponseDto = orderService.getOrder(id);
         return new ResponseEntity<>(orderResponseDto,HttpStatus.OK);
     }
@@ -39,14 +51,24 @@ public class OrderController {
     //UpdateOrderRequestDto 클래스를 매개변수로 받습니다.
     @PutMapping("/{orderId}")
     public String updateOrder(@PathVariable("orderId") Long id,
-                              @RequestBody UpdateOrderRequestDto updateOrderRequestDto) {
+                              @RequestBody UpdateOrderRequestDto updateOrderRequestDto,
+                              @AuthenticationPrincipal UserDetails userDetails) {
+        if(userDetails == null) {
+            return "로그인이 필요합니다.";
+        }
+
         orderService.updateOrder(id, updateOrderRequestDto);
         return "주문 수정";
     }
 
     // 4. 주문을 삭제하는 컨트롤러를 만듭니다. 이때 return 값은 "주문 삭제하기"입니다.
     @DeleteMapping(value = "/{orderId}")
-    public String deleteOrder(@PathVariable("orderId") Long id) {
+    public String deleteOrder(@PathVariable("orderId") Long id,
+                              @AuthenticationPrincipal UserDetails userDetails) {
+        if(userDetails == null) {
+            return "로그인이 필요합니다.";
+        }
+
         orderService.deleteOrder(id);
         return "주문 삭제하기";
     }
